requirements.txt
discord.py[voice]~=1.7.3
toml~=0.10.2
python-dotenv~=0.19.0
youtube_dl~=2021.6.6
